// src/app/features/aromas/model/aroma-type.enum.ts
export enum AromaType {
  // WOODY
  WOODY = 'WOODY', // Amaderado
  PINE = 'PINE', // Pino
  CEDAR = 'CEDAR', // Cedro
  SANDALWOOD = 'SANDALWOOD', // Sándalo
  RESINOUS = 'RESINOUS', // Resinoso
  BALSAMIC = 'BALSAMIC', // Balsámico
  OAK = 'OAK',

  // CITRUS & FRUIT
  CITRUS = 'CITRUS',  // Cítrico
  LEMON = 'LEMON', // Limón
  ORANGE = 'ORANGE', // Naranja
  GRAPEFRUIT = 'GRAPEFRUIT', // Pomelo
  LIME = 'LIME', // Lima
  TANGY = 'TANGY', // Ácido
  SOUR = 'SOUR', // Agrio
  BERRY = 'BERRY', // Frutos rojos
  GRAPE = 'GRAPE', // Uva
  APPLE = 'APPLE',  // Manzana
  PEACH = 'PEACH', // Melocotón
  APRICOT = 'APRICOT', // Albaricoque
  PLUM = 'PLUM', // Ciruela
  LYCHEE = 'LYCHEE', // Lichi
  FRUITY = 'FRUITY', // Afrutado

  // SWEET & FLORAL
  SWEET = 'SWEET', // Dulce
  HONEY = 'HONEY', // Miel
  SUGAR = 'SUGAR', // Azúcar
  CANDY = 'CANDY', // Caramelo
  AMBER = 'AMBER', // Ámbar
  FLORAL = 'FLORAL', // Floral
  ROSE = 'ROSE', // Rosa
  LAVENDER = 'LAVENDER', // Lavanda
  LILAC = 'LILAC', // Lila
  JASMINE = 'JASMINE', // Jazmín
  VIOLET = 'VIOLET', // Violeta

  // HERBAL
  HERBAL = 'HERBAL', // Herbal
  GRASSY = 'GRASSY', // Hierba fresca
  GREEN = 'GREEN', // Verde
  TEA = 'TEA', // Té
  HAY = 'HAY', // Heno
  EUCALYPTUS = 'EUCALYPTUS', // Eucalipto
  MINTY = 'MINTY', // Mentolado
  MENTHOL = 'MENTHOL', // Mentol
  SPEARMINT = 'SPEARMINT', // Hierbabuena

  // SPICY
  SPICY = 'SPICY', // Especiado
  PEPPERY = 'PEPPERY', // Pimienta
  CLOVES = 'CLOVES', // Clavo
  CINNAMON = 'CINNAMON', // Canela
  GINGER = 'GINGER', // Jengibre
  ANISE = 'ANISE', // Anís
  PUNGENT = 'PUNGENT', // Picante
  SHARP = 'SHARP', // Punxante
  ACRID = 'ACRID', // Acre

  // EARTHY & GAS
  EARTHY = 'EARTHY', // Terroso
  MUSKY = 'MUSKY', // Almizclado
  MUSTY = 'MUSTY', // Húmedo
  MOSSY = 'MOSSY', // Musgo
  PATCHOULI = 'PATCHOULI', // Pachuli
  TOBACCO = 'TOBACCO', // Tabaco
  LEATHER = 'LEATHER',  // Cuero
  FUEL = 'FUEL', // Combustible
  GAS = 'GAS', // Gas
  DIESEL = 'DIESEL', // Diésel
  KEROSENE = 'KEROSENE', // Queroseno
  CHEMICAL = 'CHEMICAL', // Químico
  SOLVENT = 'SOLVENT', // Disolvente

  // CHEESY / FUNKY
  CHEESY = 'CHEESY', // Queso
  SKUNK = 'SKUNK', // Skunk
  SULFUROUS = 'SULFUROUS', // Sulfurado
  FUNKY = 'FUNKY', // Funky
  STINKY = 'STINKY', // Apestoso
  ROTTEN = 'ROTTEN', // Podrido
  SWEATY = 'SWEATY', // Sudor
  MEDICINAL = 'MEDICINAL', // Medicinal
  CAMPHOR = 'CAMPHOR', // Alcanfor
  WAXY = 'WAXY', // Cera
  OILY = 'OILY', // Aceitoso
  MUSTARD = 'MUSTARD' // Mostaza
}